Reorganize into header and namespace
